Perform better cleanup of DNS/routing settings
Sometimes when disconnected the VPN-specific DNS or routing
settings still remain active. When switching to vpnc for cleanup
the vpn gateway and tun device settings were not stored,
so an extra cleanup step was not performed.

Letting vpnc perform cleanup on disconnect, and on reconnect
we run our extra cleanup which tries to remove erroneous
DNS settings.